test: move test from FormField to Form
